⚡ Bolt: Replace double precision math functions with float precision ones

+++ src/ZoneManager.h
@@ -104,7 +104,7 @@ public:
         long minDistSq = (long)z.minDist * z.minDist;
         long maxDistSq = (long)z.maxDist * z.maxDist;
         if (distSq < minDistSq || distSq > maxDistSq) return false;
-        float angleRad = atan2((float)x, (float)y);
+        float angleRad = atan2f((float)x, (float)y);
         int angleDeg = (int)(angleRad * 180.0f / PI);
         if (angleDeg < z.minAngle || angleDeg > z.maxAngle) return false;
         return true;
